fix(workspace): expand leading ~ in user-supplied project paths
A hand-typed location in the New Project screen (placeholder "~/Projects")
was passed through verbatim. The filesystem never expands ~, so
create_empty_repo / git clone built a path with a literal ~, created a
phantom directory named "~" under the process cwd, and the workspace
stored "~/..." as its cwd. Every terminal then failed to chdir into that
path and fell back to $HOME.

Promote expand_tilde into the shared project module (previously private
to the PTY daemon) and apply it at the path-entry commands:
check_is_git_repo, init_git_repo, create_empty_repo, git_clone_repo,
create_workspace, and create_empty_workspace.
